:sparkles: Implement mouse based Rotation
Implements a simple version of mouse based rotation.Adds a function in
NetworkManagerClient to get to obtain the local player postion, so the
rotation angle can be calculated. Changes the network serilisation so
the new rotation is sent as an update to all players. Adds a getter to
the RenderManager to allow access to the camera view for calculating the
mouse postion relative to world coordinates.